+++ src/main/java/com/ecommerce/bicicle/repository/ItemImgUrlsRepository.java
@@ -3,6 +3,6 @@ package com.ecommerce.bicicle.repository;
 import com.ecommerce.bicicle.entity.ItemImgUrls;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemImgUrlsRepository extends CrudRepository<ItemImgUrls, Long> {
+public interface ItemImgUrlsRepository extends CrudRepository<ItemImgUrls, Integer> {
 
 }
